requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
aom=3.6.0=h6a678d5_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
blas=1.0=mkl
blosc=1.21.3=h6a678d5_0
bottleneck=1.3.7=py38ha9d4c09_0
brotli=1.0.9=h5eee18b_8
brotli-bin=1.0.9=h5eee18b_8
brotli-python=1.0.9=py38h6a678d5_8
brunsli=0.1=h2531618_0
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=py38h06a4308_0
cfitsio=3.470=h5893167_7
chardet=5.2.0=pypi_0
charls=2.2.0=h2531618_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py38h06a4308_0
cloudpickle=2.2.1=py38h06a4308_0
colour=0.1.5=pypi_0
contourpy=1.1.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.4=hbda6634_3
cycler=0.12.1=pypi_0
cytoolz=0.12.2=py38h5eee18b_0
dask-core=2023.4.1=py38h06a4308_0
dav1d=1.2.1=h5eee18b_0
easydict=1.13=pypi_0
fightingcv-attention=1.0.0=pypi_0
filelock=3.13.1=py38h06a4308_0
fonttools=4.47.2=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
google-auth=2.27.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.60.0=pypi_0
huggingface-hub=0.23.0=pypi_0
idna=3.7=py38h06a4308_0
imagecodecs=2023.1.23=py38hc4b7b5f_0
imageio=2.33.1=py38h06a4308_0
importlib-metadata=7.0.1=py38h06a4308_0
importlib_resources=6.1.1=py38h06a4308_1
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py38h06a4308_0
jpeg=9e=h5eee18b_1
jxrlib=1.1=h7b6447c_2
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h143b758_1
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libaec=1.0.4=he6710b0_1
libavif=0.11.1=h5eee18b_0
libbrotlicommon=1.0.9=h5eee18b_8
libbrotlidec=1.0.9=h5eee18b_8
libbrotlienc=1.0.9=h5eee18b_8
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.8.1.2=0
libcurand=10.3.5.147=h99ab3db_1
libcurl=8.5.0=h251f7ec_0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libnghttp2=1.57.0=h2d74bed_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libssh2=1.10.0=hdbd6064_3
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.1=h6a678d5_0
libwebp-base=1.3.2=h5eee18b_0
libzopfli=1.0.3=he6710b0_0
lightning=2.1.3=pypi_0
lightning-utilities=0.10.1=pypi_0
linear-warmup-cosine-annealing-warm-restarts-weight-decay=1.0=pypi_0
locket=1.0.0=py38h06a4308_0
lz4-c=1.9.4=h6a678d5_1
markdown=3.5.2=pypi_0
markupsafe=2.1.3=py38h5eee18b_0
matplotlib=3.7.4=pypi_0
matplotlib-base=3.7.2=py38h1128e8f_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=py38h06a4308_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
oauthlib=3.2.2=pypi_0
opencv-python=4.9.0.80=pypi_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_1
packaging=23.2=py38h06a4308_0
pandas=2.0.3=py38h1128e8f_0
partd=1.4.1=py38h06a4308_0
pillow=10.3.0=py38h5eee18b_0
pip=23.3.1=py38h06a4308_0
platformdirs=3.10.0=py38h06a4308_0
pooch=1.7.0=py38h06a4308_0
protobuf=4.25.2=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=2.0.1=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-lightning=2.1.3=pypi_0
pytorch-mutex=1.0=cuda
pytz=2024.1=py38h06a4308_0
pywavelets=1.4.1=py38h5eee18b_0
pyyaml=6.0.1=py38h5eee18b_0
readline=8.2=h5eee18b_0
requests=2.31.0=py38h06a4308_1
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scikit-image=0.19.2=py38h43a58ef_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=py38h06a4308_0
setuptools=68.2.2=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.10=h6a678d5_1
sqlite=3.45.3=h5eee18b_0
sympy=1.12=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
tifffile=2023.4.12=py38h06a4308_0
timm=0.9.16=pypi_0
tk=0.1.0=pypi_0
tkmacosx=1.0.5=pypi_0
toolz=0.12.0=py38h06a4308_0
torchaudio=2.0.2=py38_cu118
torchmetrics=1.3.0.post0=pypi_0
torchtriton=2.0.0=py38
torchvision=0.15.2=cpu_py38h83e0c9b_0
tqdm=4.66.1=pypi_0
ttkwidgets=0.13.0=pypi_0
typing_extensions=4.9.0=py38h06a4308_1
tzdata=2024.1=pypi_0
unicodedata2=15.1.0=py38h5eee18b_0
urllib3=2.1.0=py38h06a4308_1
werkzeug=3.0.1=pypi_0
wheel=0.43.0=py38h06a4308_0
xz=5.4.6=h5eee18b_1
yaml=0.2.5=h7b6447c_0
yarl=1.9.4=pypi_0
zfp=1.0.0=h6a678d5_0
zipp=3.17.0=py38h06a4308_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
